refactor(security): 重构数据安全工具类并完善软删除配置
- 配置MyBatis-Plus逻辑删除字段和值
- 将DataSecurityUtil改为单例模式并添加INSTANCE实例
- 移除无用的SymmetricCrypto和Slf4j注解导入
- 新增CODE_ERROR_OR_EXPIRED错误码枚举
- 将JWT过期时间从Integer改为Long类型
- 添加BouncyCastle依赖用于SM4加密支持
- TypeHandler中直接使用DataSecurityUtil.INSTANCE替换注入方式
- 优化AuthContext获取用户ID的安全方法
- 修正数据安全属性配置前缀
- 字段填充处理器使用OffsetDateTime替换LocalDateTime
- 移动WebMvcConfig包路径并排除Swagger接口文档路径
- 统一依赖管理中的BouncyCastle版本配置

+++ common/src/main/java/interview/common/util/JwttUtil.java
@@ -20,11 +20,11 @@ import java.util.Optional;
 @Slf4j
 public class JwttUtil {
 
-    private final Integer expiration;
+    private final Long expiration;
     private final SecretKey key;
     private final JwtParser parser;
 
-    public JwttUtil(Integer expiration, String secretString) {
+    public JwttUtil(Long expiration, String secretString) {
         this.expiration = expiration;
         byte[] bytes = secretString.getBytes(StandardCharsets.UTF_8);
         this.key = Keys.hmacShaKeyFor(bytes);
@@ -45,7 +45,7 @@ public class JwttUtil {
 
         long nowMillis = System.currentTimeMillis();
         Date now = new Date(nowMillis);
-        Date expireDate = new Date(nowMillis + (expiration.longValue() * 60 * 1000));
+        Date expireDate = new Date(nowMillis + (expiration * 60 * 1000));
         return Jwts.builder()
                 .claims(data != null ? data : Map.of())
                 .subject(userIdString)
